Create Add_AD_Description_to_Notes.ps1
Adds AD Description to Notes
This version of the script requires Quest Active Roles management plugin or can be modified to use MS AD module. 
This script will query Active Directory and find the Active Directory description for each VM, 
then write the description to the Notes property of the VM. If a computer account does not
exist in AD it will leave the notes blank.